# openapi.yaml
openapi: 3.1.0
info:
  title: AgentAuth API
  version: 1.0.0
  description: |
    Authorization layer for AI agent commerce.
    AgentAuth enables developers to set policies, budgets, and approval flows
    for AI agents making purchases on behalf of users.
  contact:
    name: AgentAuth
    url: https://agentauth.in
  license:
    name: Proprietary
servers:
  - url: https://agentauth-api.koyeb.app
    description: Production
  - url: http://localhost:8000
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    AuthorizeRequest:
      type: object
      required: [agent_id, user_id, action, amount]
      properties:
        agent_id:
          type: string
          example: "agent_shopping_bot"
        user_id:
          type: string
          example: "user_abc123"
        action:
          type: string
          example: "purchase"
        amount:
          type: number
          example: 49.99
        currency:
          type: string
          default: "USD"
        merchant_id:
          type: string
          example: "merch_walmart"
        merchant_name:
          type: string
          example: "Walmart"
        category:
          type: string
          example: "groceries"
        metadata:
          type: object
          additionalProperties: true
    AuthorizeResponse:
      type: object
      properties:
        status:
          type: string
          enum: [approved, denied, requires_approval, rate_limited, error]
        authorized:
          type: boolean
        request_id:
          type: string
        token:
          type: string
          nullable: true
        reason:
          type: string
        risk_score:
          type: number
        policy_id:
          type: string
          nullable: true
        constraints:
          type: object
        evaluation_time_ms:
          type: number

    # ── Consent ──
...
    Error:
      type: object
      properties:
        detail:
          type: string
paths:
...
  /v1/authorize:
    post:
      tags: [Authorization]
      summary: Authorize an agent action
      description: Evaluate policies and authorize (or deny) an agent's requested action.
      operationId: authorize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizeRequest"
      responses:
        "200":
          description: Authorization decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizeResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/billing/checkout:
    post:
      tags: [Billing]
      summary: Create checkout session
      operationId: createCheckoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan, success_url, cancel_url]
              properties:
                plan:
                  type: string
                  enum: [startup, growth, enterprise]
                success_url:
                  type: string
                cancel_url:
                  type: string
      responses:
        "200":
          description: Checkout URL
  /v1/billing/portal:
    get:
      tags: [Billing]
      summary: Get billing portal URL
      operationId: getBillingPortal
      parameters:
        - name: return_url
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Portal URL

  # ── API Keys ──
...